Bug 1422898 - Update Harfbuzz to 1.7.2. r=jfkthame
UltraBlame original commit: d5e064a4bca13180d70734ac83543c3a29b47a4f

+++ gfx/harfbuzz/src/hb-open-file-private.hh
@@ -64,9 +64,9 @@ typedef struct TableRecord
 
   Tag		tag;		
   CheckSum	checkSum;	
-  ULONG		offset;		
+  UINT32		offset;		
 
-  ULONG		length;		
+  UINT32		length;		
   public:
   DEFINE_SIZE_STATIC (16);
 } OpenTypeTable;
@@ -154,7 +154,7 @@ struct TTCHeaderVersion1
   Tag		ttcTag;		
   FixedVersion<>version;	
 
-  ArrayOf<LOffsetTo<OffsetTable>, ULONG>
+  ArrayOf<LOffsetTo<OffsetTable>, UINT32>
 		table;		
 
   public:
